+++ packages/cql/lib/engine/expression/operator_expressions/unary_expressions/predecessor.dart
@@ -138,6 +138,7 @@ class Predecessor extends UnaryExpression {
         return value - ExtendedDuration(months: 1);
       }
       if (value.daysPrecision) {
+        print('result: ${value - ExtendedDuration(days: 1)}');
         return value - ExtendedDuration(days: 1);
       }